backend/requirements.txt
wheel
django==3.1.8
dj-stripe
django-rest-framework
django-cors-headers
django-filter
markdown
pygithub
sentry-sdk
python-memcached
django-q
